create useCategoryPosts hook
